Move CategoryViewSet from api app to categories app
