+++ project/api/views/__init__.py
@@ -1,4 +1,2 @@
-from accounts.api import ProfileViewSet
-from .category import CategoryViewSet
 from .civi import CiviViewSet
 from .thread import ThreadViewSet
